extract metaproramm-y stuff into the extend module
